Extract CompletedUpgradeView
